Add a little jitter to the peering keepalives
This should distribute the peering keepalives evenly after some time of functioning. There is already such an implicit jitter — the `PATCH` operation with varying duration (due to its networking nature). But it is not controllable, and the distribution in time happens slowly. With multiple namespaces served (each with its own peering keepalive), it would be better to distribute them faster to reduce the peak load on the K8s API.