Resume from freeze immediately when blocking peers expire
Previously, the unfreezing only happened on the next occurrence when the peering resource was updated. Since most blocking peers were gone by that time, the update only happened from the operator's own keepalive change. Which lead to operators not unfreezing for up to 1 minute by default.

With this change, the operators will estimate how long it is left till the blocking peers are supposed to expire, and —if there were no other signals or activities from other peers— will unfreeze immediately.

The unfreeze is triggered by a fake update on the peering resource for the same operator — i.e. the same as keepalive would do in he next 0..60 seconds. But this is not a problem in case the peering was completely idle for some time. This also ensures that the unfreeze will be re-evaluated by the actual state of the peering resource, not an in-memory assumption about it.